Part 2 of functional specification: Valuations!
This is part 2 of the functional specification and an update to the database description for the new "tables" that are related to valuations. Getting the specification out quickly was more important to me than a careful review. I will do that later if need be. Feel free to ask me for clarifications or better wording as soon as any questions come up. Thank you.